Move test of the gene information merger from core to the merger directory
